Feature: Havok Collision Gen in Blender (WIP)
This is based on the "import obj as collision" with NifSkope Dev11.

This contains a class `HavokCollisionGenerator` to create one/more
empties depending on user selection that will contain Havok collision
data.

If a user has multiple meshes selected named `*collision*` and a single
non-`*collision*` mesh selected, they will all be applied to that mesh.

If a user has a single `*collision*` mesh selected and multiple
non-`*collision*` meshes selected, they'll all get the same collision
data applied.

Also adds an n-panel named "PyNifly" to hold the button to create the
data.

The Havok collision data is initially created with these defaults:
- `Collision Layer`: STATIC (1)
- `Collision Filter Bitmask`: `0xFA859241` (borrowed from NifSkope)
- `Material Selection`: `STONE` (Enum value from existing PyNifly code)
- `Mass`: 0.0
- `Friction`: 0.5
- `Restitution`: 0.4 (Bounciness)
- `Motion System`: 7 (Fixed, rigid body/kinematic)
- `Quality Type`: 1
- `Solver Deactivation`: 1 (Inverse of how booleans normally work...)

When activating, an empty is created containing the above settings as
custom properties with the same center as the target object (or object
in the loop, depending on what the user chose).

This creates a hierarchy like:

```
Sword (Target Mesh)
└───  Collision_Sword (Empty with Havok properties)
    ├── Shape_Collision_A (Child mesh)
    └── Shape_Collision_B (Child mesh)

```